pyproject.toml
[tool.poetry]
name = "polling"
version = "0.1.0"
description = "Polling API"
authors = ["captainKoala <konstantin.paulkin@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "2.2.10"
python-dotenv = "^0.19.0"
djangorestframework = "^3.12.4"
gunicorn = "^20.1.0"
psycopg2-binary = "2.8.6"
PyYAML = "^5.4.1"
uritemplate = "^3.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
